+++ app/dashboard/index.js
@@ -107,7 +107,6 @@ dashboard.post(
 
 // Account page does not need to know about the state of the folder
 // for a particular blog
-
 dashboard.use(function (req, res, next) {
   res.locals.links_for_footer = [];
   res.locals.partials = res.locals.partials || {};
@@ -131,13 +130,9 @@ dashboard.use("/status", require("./routes/status"));
 
 // Special function which wraps render
 // so there is a default layout and a partial
-// inserted into it
 dashboard.use(require("./render"));
 
-dashboard.use(function (req, res, next) {
-  res.locals.breadcrumbs = new Breadcrumbs();
-  next();
-});
+dashboard.use(require("./breadcrumbs"));
 
 dashboard.use("/account", require("./routes/account"));
 
@@ -163,7 +158,7 @@ dashboard.use("/dashboard/:handle", function (req, res, next) {
 // Load the files and folders inside a blog's folder
 dashboard.get(
   "/dashboard/:handle/folder/:path*",
-  
+
   function (req, res, next) {
     req.folderPath = "/" + req.params.path;
     next();
@@ -178,31 +173,10 @@ dashboard.get(
 
 dashboard.get("/dashboard/:handle", require("./routes/folder"));
 
-function Breadcrumbs() {
-  var list = [];
-
-  list.add = function (label, slug) {
-    var base = "/";
-
-    if (list.length) base = list[list.length - 1].url;
-
-    list.push({ label: label, url: require("path").join(base, slug) });
-
-    for (var i = 0; i < list.length; i++) {
-      list[i].first = i === 0;
-      list[i].last = i === list.length - 1;
-      list[i].only = i === 0 && list.length === 1;
-    }
-  };
-
-  return list;
-}
-
 dashboard.use("/dashboard/:handle", require("./routes/settings"));
 
-dashboard.use(require("./routes/settings/errorHandler"));
-
 // need to handle dashboard errors better...
+dashboard.use(require("./routes/settings/errorHandler"));
 dashboard.use(require("./routes/error"));
 
 // Restore render function, remove this dumb bullshit eventually
