+++ app/dashboard/views/account/change-password.html
@@ -17,7 +17,7 @@
 
   <div class="buttons">
     <button type="submit" onclick="this.form.submit();this.disabled=true">Save new password</button>
-    <a href="/">Cancel</a>
+    <a href="/dashboard">Cancel</a>
   </div>
   <br><br>
 </form>
