+++ app/dashboard/views/account/email.html
@@ -7,7 +7,7 @@
 
   <div class="buttons">
     <button type="submit" onclick="this.form.submit();this.disabled=true">Save email address</button>
-    <a href="/">Cancel</a>
+    <a href="/dashboard">Cancel</a>
   </div>
 
   <br><br><br>
