+++ app/dashboard/views/settings/date.html
@@ -1,4 +1,3 @@
-<input type="hidden" name="redirect" value="/settings/services/date">
 
 <style type="text/css">
   form label {
