+++ app/dashboard/views/settings/link-format.html
@@ -13,8 +13,6 @@
   .format:hover{background: rgba(0,0,0,.02)}
 </style>
 
-<input type="hidden" name="redirect" value="/settings/services/link-format">
-
 <p>
   Each post is assigned a URL according to the format below. You can override a post's URL in its <a href="/how/metadata">metadata</a>. If you change your URL format, Blot redirects the old URLs automatically.
 </p>
