+++ app/dashboard/views/settings/redirects.html
@@ -1,4 +1,3 @@
-<input type="hidden" name="redirect" value="/settings/services/redirects">
 <input type="hidden" name="has_redirects" value="true">
 
 <p>
