+++ app/dashboard/views/settings/services.html
@@ -1,13 +1,13 @@
 <style type="text/css">form {max-width:none}</style>
 <div style="margin:-11px -26px">
 
-  <a class="line" href="/settings/services/date">
+  <a class="line" href="{{{base}}}/services/date">
     <span class="label">Date and time</span>
     <span class="center">{{currentTime}}</span>
     <span class="right" >Edit</span>    
   </a>
 
-  <a class="line" href="/settings/services/link-format">
+  <a class="line" href="{{{base}}}/services/link-format">
     <span class="label">Link format</span>
     <span class="center">
       {{#permalinkFormats}}
@@ -17,7 +17,7 @@
   </a>
 
 
-  <a class="line" href="/settings/services/redirects">
+  <a class="line" href="{{{base}}}/services/redirects">
     <span class="label">Redirects</span>
     <span class="link">Edit your redirect rules</span>
   </a>
